Assignment4.0-index.html
I converted the weight and temperature values in the story from pounds and Fahrenheit into stones and centigrade

I searched up the formulas for converting pounds to stone, and Fahrenheit to centigrade.
Inside the line that defines the weight variable and replaced the 300 with a calculation that converts 300 pounds into stones. and  Concatenated the ' stone' onto the end of the result of the overall Math.round() call.
Inside the line that defines the temperature variable, replace 94 with a calculation that converts 94 Fahrenheit into centigrade. Concatenate ' centigrade' onto the end of the result of the overall Math.round() call.
Just under the two variable definitions, add two more string replacement lines that replace '94 fahrenheit' with the contents of the temperature variable, and '300 pounds' with the contents of the weight variable.